requirements.txt
Flask
Flask-WTF
requests
pymysql
argon2-cffi
WTForms
Flask-Limiter
redis
email-validator
APScheduler
cryptography
Flask-Session
flask_caching
pillow